TODO 2018/08/05

+++ src/main/java/com/ilongli/service/UserService.java
@@ -7,5 +7,5 @@ import com.ilongli.entity.User;
  * @author ilongli
  *
  */
-public interface UserService extends BaseService<User, Integer> {
+public interface UserService extends BaseService<User, Long> {
 }
